Move strategies out.
